Many changes
A lot of functions were removed, but will be slowly re-implanted. 
There was an error with the NULL detection of pointers, as we needed the arduino.h library.
I was also stuck in changing the value of a local pointer in a function which called another function which was suppose to change the pointer. but it changed a pointer, which seemed to disappear after the end of the function. So I put the rest of the function which called other functions using the pointer in my  other function. Which seemed to fix the problem. (function which declares pointers -> calls function which changes pointers -> returns int and should have changed  the pointers -> initial function uses the pointers in other called function, but they seem to be gone) => (function -> calls function which creates pointers -> calls functions which uses the pointer-> returnx2). Very sorry if I wasn't clear